Hexadecimal reception
Added the possibility to display received data as hexadecimal values.
This feature also works when timestamps are enabled as long as the data
is terminated by a linefeed character.

Also, some cleanup of code was performed. Printing data to the received
data text box is now performed by a dedicated method.